Remove dead code, and other cleanups
Of note: The override of fillStackTrace in ReadPastEndException has been removed, because it's only initialised once, and its presence makes Class Hierarchy Analysis nervous about inlining it.
