+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/IllegalResourceException.java
@@ -21,23 +21,8 @@ package org.neo4j.kernel.impl.transaction;
 
 public class IllegalResourceException extends RuntimeException
 {
-    public IllegalResourceException()
-    {
-        super();
-    }
-
     public IllegalResourceException( String message )
     {
         super( message );
-    }
-
-    public IllegalResourceException( String message, Throwable cause )
-    {
-        super( message, cause );
-    }
-
-    public IllegalResourceException( Throwable cause )
-    {
-        super( cause );
     }
 }
