+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/log/ReadPastEndException.java
@@ -34,10 +34,4 @@ public final class ReadPastEndException extends IOException
     private ReadPastEndException()
     {
     }
-
-    @Override
-    public Throwable fillInStackTrace()
-    {
-        return this;
-    }
 }
