+++ community/kernel/src/test/java/org/neo4j/kernel/impl/transaction/log/entry/LogEntryConsumerTest.java
@@ -24,10 +24,6 @@ import org.junit.Test;
 import org.neo4j.helpers.Function;
 import org.neo4j.kernel.impl.transaction.command.LogHandler;
 import org.neo4j.kernel.impl.transaction.log.LogPosition;
-import org.neo4j.kernel.impl.transaction.log.entry.LogEntryCommand;
-import org.neo4j.kernel.impl.transaction.log.entry.LogEntryStart;
-import org.neo4j.kernel.impl.transaction.log.entry.OnePhaseCommit;
-import org.neo4j.kernel.impl.transaction.log.entry.TranslatingEntryVisitor;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
