+++ community/kernel/src/test/java/org/neo4j/kernel/impl/transaction/log/pruning/TransactionCountThresholdTest.java
@@ -19,19 +19,18 @@
  */
 package org.neo4j.kernel.impl.transaction.log.pruning;
 
+import org.junit.Test;
+
+import java.io.File;
+
+import org.neo4j.kernel.impl.transaction.log.IllegalLogFormatException;
+import org.neo4j.kernel.impl.transaction.log.LogFileInformation;
+
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-
-import java.io.File;
-
-import org.junit.Test;
-
-import org.neo4j.kernel.impl.transaction.log.IllegalLogFormatException;
-import org.neo4j.kernel.impl.transaction.log.LogFileInformation;
-import org.neo4j.kernel.impl.transaction.log.pruning.TransactionCountThreshold;
 
 public class TransactionCountThresholdTest
 {
